adjust the definition of major_minor
Former-commit-id: 25d1e64ef714bbac00c9e21d8fcf37fdbd2a7092

+++ libstreetmap/src/m2_more.h
@@ -81,7 +81,7 @@ float lon_from_x(double x);
 struct action_mem{
     // store intersections u clickded
     int last_clicked_intersection=-1;
-    int initial_world_width=0;
+    float initial_world_width=0;
 };
 
 action_mem memory;
